fix: Code cleanup and OR clause fix
- Separate code into water body management utils file
- api file must be minimal
- In OR clause use braces to separate operands to give internal OR effect not global OR
